+++ ChangeLog.md
@@ -3,3 +3,13 @@
 ## 1.0  -- 2018-05-11
 
 * First version. Released on an unsuspecting world.
+
+## 1.1  -- 2018-05-27
+
+### Added
+- Bottleneck distance, a way to compare the topology of different data sets.
+- HasseDiagram module, will allow users to deduce topological features of information flow in networks.
+- A function for encoding generic graphs as 2D vectors.
+
+### Changed
+- Improved documentation for all exposed modules.